+++ src/django_resume/views.py
@@ -1,5 +1,10 @@
+from typing import Any
+
+from django.contrib.auth.decorators import login_required
+from django import forms
 from django.http import HttpRequest, HttpResponse
 from django.shortcuts import render, get_object_or_404
+from django.views.decorators.http import require_http_methods
 
 from .models import Resume
 from .plugins import plugin_registry
@@ -84,12 +89,54 @@ def resume_detail(request: HttpRequest, slug: str) -> HttpResponse:
     return render(request, "django_resume/plain/resume_detail.html", context=context)
 
 
+class ResumeForm(forms.ModelForm):
+    class Meta:
+        model = Resume
+        fields = ["name", "slug"]
+
+
+@login_required
+@require_http_methods(["GET", "POST"])
 def resume_list(request: HttpRequest) -> HttpResponse:
     """
-    The main resume list view.
+    The main resume list view. Only authenticated users can see it.
 
-    You can add and delete resumes from this view.
+    You can add and delete your resumes from this view.
     """
-    context = {"resumes": Resume.objects.all()}
-    print("context: ", context)
-    return render(request, "django_resume/plain/resume_list.html", context=context)
+    assert request.user.is_authenticated  # type guard just to make mypy happy
+    my_resumes = Resume.objects.filter(owner=request.user)
+    context: dict[str, Any] = {
+        "is_editable": True,  # needed to include edit styles in the base
+        "resumes": my_resumes,
+        "form": ResumeForm(),
+    }
+    if request.method == "POST":
+        form = ResumeForm(request.POST)
+        context["form"] = form
+        if form.is_valid():
+            resume = form.save(commit=False)
+            resume.owner = request.user
+            resume.save()
+            context["new_resume"] = resume
+        return render(
+            request, "django_resume/plain/resume_list_main.html", context=context
+        )
+    else:
+        # just render the complete template on GET
+        return render(request, "django_resume/plain/resume_list.html", context=context)
+
+
+@login_required
+@require_http_methods(["DELETE"])
+def resume_delete(request: HttpRequest, slug: str) -> HttpResponse:
+    """
+    Delete a resume.
+
+    Only the owner of the resume can delete it.
+    """
+    resume = get_object_or_404(Resume, slug=slug)
+    if resume.owner != request.user:
+        return HttpResponse(status=403)
+
+    resume.delete()
+    return HttpResponse(status=200)  # 200 instead of 204 for htmx compatibility